requirements.txt
pydantic~=1.10.4
fastapi~=0.90.0
requests~=2.28.2
uvicorn~=0.20.0
python-dotenv~=0.21.1